Rename the verify function into verifyLicense.
