+++ packages/ckeditor5-utils/src/verifylicense.ts
@@ -4,7 +4,7 @@
  */
 
 /**
- * @module utils/madewith
+ * @module utils/verifylicense
  */
 
 import { releaseDate } from './version';
@@ -20,7 +20,7 @@ export type VerifiedKeyStatus = 'VALID' | 'INVALID';
  * @param token The string to check.
  * @returns String that represents the state of given `token` parameter.
  */
-export default function verify( token: string | undefined ): VerifiedKeyStatus {
+export default function verifyLicense( token: string | undefined ): VerifiedKeyStatus {
 	// TODO: issue ci#3175
 	let decryptedData = '';
 	let decryptedSecondElement = '';
