+++ packages/ckeditor5-utils/tests/verifylicense.js
@@ -3,7 +3,7 @@
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
-import verify from '../src/madewith';
+import verifyLicense from '../src/verifylicense';
 
 describe( 'utils', () => {
 	describe( 'verify', () => {
@@ -13,7 +13,7 @@ describe( 'utils', () => {
 				// eslint-disable-next-line max-len
 				const string = 'dG9vZWFzZXRtcHNsaXVyb3JsbWlkbXRvb2Vhc2V0bXBzbGl1cm9ybG1pZG10b29lYXNldG1wc2xpdXJvcmxtaWRtLU1qQTBOREEyTVRJPQ==';
 
-				expect( verify( string ) ).to.be.equal( 'VALID' );
+				expect( verifyLicense( string ) ).to.be.equal( 'VALID' );
 			} );
 
 			it( 'when old token format is given', () => {
@@ -21,33 +21,33 @@ describe( 'utils', () => {
 				// eslint-disable-next-line max-len
 				const string = 'YWZvb2JkcnphYXJhZWJvb290em9wbWJvbHJ1c21sZnJlYmFzdG1paXN1cm1tZmllenJhb2F0YmFvcmxvb3B6aWJvYWRiZWZzYXJ0bW9ibG8=';
 
-				expect( verify( string ) ).to.be.equal( 'VALID' );
+				expect( verifyLicense( string ) ).to.be.equal( 'VALID' );
 			} );
 		} );
 
 		describe( 'should return `INVALID`', () => {
 			it( 'when token is empty', () => {
-				expect( verify( '' ) ).to.be.equal( 'INVALID' );
+				expect( verifyLicense( '' ) ).to.be.equal( 'INVALID' );
 			} );
 
 			it( 'when token is not passed', () => {
-				expect( verify( ) ).to.be.equal( 'INVALID' );
+				expect( verifyLicense( ) ).to.be.equal( 'INVALID' );
 			} );
 
 			describe( 'new', () => {
 				it( 'first too short', () => {
-					expect( verify( 'Wm05dlltRnktTWpBeU5UQXhNREU9' ) ).to.be.equal( 'INVALID' );
+					expect( verifyLicense( 'Wm05dlltRnktTWpBeU5UQXhNREU9' ) ).to.be.equal( 'INVALID' );
 				} );
 
 				it( 'first too long', () => {
 					// eslint-disable-next-line max-len
 					const string = 'YzNSbGJTQmxjbkp2Y2pvZ2JtVjBPanBGVWxKZlFreFBRMHRGUkY5Q1dWOURURWxGVGxSemRHVnRJR1Z5Y205eU9pQnVaWFE2T2tWU1VsOUNURTlEUzBWRVgwSlpYME5NU1VWT1ZITjBaVzBnWlhKeWIzSTZJRzVsZERvNlJWSlNYMEpNVDBOTFJVUmZRbGxmUTB4SlJVNVVjM1JsYlNCbGNuSnZjam9nYm1WME9qcEZVbEpmUWt4UFEwdEZSRjlDV1Y5RFRFbEZUbFJ6ZEdWdElHVnljbTl5T2lCdVpYUTZPa1ZTVWw5Q1RFOURTMFZFWDBKWlgwTk1TVVZPVkhOMFpXMGdaWEp5YjNJNklHNWxkRG82UlZKU1gwSk1UME5MUlVSZlFsbGZRMHhKUlU1VS1NakF5TlRBeE1ERT0=';
-					expect( verify( string ) ).to.be.equal( 'INVALID' );
+					expect( verifyLicense( string ) ).to.be.equal( 'INVALID' );
 				} );
 
 				it( 'first wrong format', () => {
 					const string = 'YS1NakF5TlRBeE1ERT0=';
-					expect( verify( string ) ).to.be.equal( 'INVALID' );
+					expect( verifyLicense( string ) ).to.be.equal( 'INVALID' );
 				} );
 
 				it( 'when date is invalid', () => {
@@ -55,20 +55,20 @@ describe( 'utils', () => {
 
 					const string = 'enN6YXJ0YWxhYWZsaWViYnRvcnVpb3Jvb3BzYmVkYW9tcm1iZm9vbS1NVGs1TnpFeA==';
 
-					expect( verify( string ) ).to.be.equal( 'INVALID' );
+					expect( verifyLicense( string ) ).to.be.equal( 'INVALID' );
 				} );
 
 				it( 'wrong second part format', () => {
 					const string = 'Ylc5MGIyeDBiMkZ5YzJsbGMybHNjR1Z0Y20xa2RXMXZkRzlzZEc5aGNuTnBaWE09LVptOXZZbUZ5WW1FPQ==';
 
-					expect( verify( string ) ).to.be.equal( 'INVALID' );
+					expect( verifyLicense( string ) ).to.be.equal( 'INVALID' );
 				} );
 
 				it( 'when wrong string passed', () => {
 					// # instead of second part
 					const string = 'Ylc5MGIyeDBiMkZ5YzJsbGMybHNjR1Z0Y20xa2RXMXZkRzlzZEc5aGNuTnBaWE09LSM=';
 
-					expect( verify( string ) ).to.be.equal( 'INVALID' );
+					expect( verifyLicense( string ) ).to.be.equal( 'INVALID' );
 				} );
 
 				it( 'when date is earlier than the release date', () => {
@@ -76,7 +76,7 @@ describe( 'utils', () => {
 					// eslint-disable-next-line max-len
 					const string = 'Y0dWc1lYVmxhWE4wYlc5a2MyMXBiRzEwY205eWIzQmxiR0YxWldsemRHMXZaSE50YVd4dGRISnZjbTlrYzJGa2MyRmtjMkU9LU1UazRNREF4TURFPQ==';
 
-					expect( verify( string ) ).to.be.equal( 'INVALID' );
+					expect( verifyLicense( string ) ).to.be.equal( 'INVALID' );
 				} );
 			} );
 
@@ -84,7 +84,7 @@ describe( 'utils', () => {
 				it( 'when date is missing', () => {
 					const string = 'dG1wb3Rhc2llc3VlbHJtZHJvbWlsbw==';
 
-					expect( verify( string ) ).to.be.equal( 'INVALID' );
+					expect( verifyLicense( string ) ).to.be.equal( 'INVALID' );
 				} );
 			} );
 
@@ -92,7 +92,7 @@ describe( 'utils', () => {
 				// invalid
 				const string = 'foobarbaz';
 
-				expect( verify( string ) ).to.be.equal( 'INVALID' );
+				expect( verifyLicense( string ) ).to.be.equal( 'INVALID' );
 			} );
 		} );
 	} );
